re-enable tower building in BuildManager
