+++ Assets/ScriptableObjects/Towers/Ballista.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b4eefd49f8a61c149988a8afad4e2491
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
