+++ Assets/ScriptableObjects/Towers/Ballista/BallistaBase_0.asset
@@ -1,0 +1,24 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 02acc0fa88f06934f9c8db381de21961, type: 3}
+  m_Name: BallistaBase_0
+  m_EditorClassIdentifier: 
+  towerName: Ballista
+  range: 3
+  versionNum: 0
+  fireRate: 1
+  element: {fileID: 0}
+  cost: 0
+  towerPrefab: {fileID: 0}
+  bulletPrefab: {fileID: 0}
+  nextUpgrade: {fileID: 0}
+  nextElements: []
