+++ Assets/ScriptableObjects/Towers/Ballista/BallistaBase_0.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 06b49e2a01d1e1e4f97cf6bb361fabe1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
