+++ Assets/ScriptableObjects/Towers/Cannon/CannonBase_0.asset
@@ -20,7 +20,7 @@ MonoBehaviour:
   cost: 30
   towerPrefab: {fileID: 0}
   bulletPrefab: {fileID: 0}
-  nextUpgrade: {fileID: 11400000, guid: 8a87d74afed0e474190f5d0feef49ef1, type: 2}
+  nextUpgrade: {fileID: 11400000, guid: fbe7e6c9477214940a6e03d2fb9952d9, type: 2}
   nextElements:
   - element: {fileID: 0}
     elementTowerInfo: {fileID: 11400000, guid: cfae3c27644fb27438f3d4c8e8b83188, type: 2}
