+++ Assets/ScriptableObjects/Towers/Spikey.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 33ae16c836b3ca6478d1a9675b10c4e2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
