+++ Assets/ScriptableObjects/Towers/Spikey/SpikeyBase_0.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7e608828053941e4580916f257cb5fa0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
